requirements.txt
swisseph
pandas
numpy
yfinance
yahoofinancials
datetime
jupyter
ipykernel
notebook
jupyterlab
matplotlib
seaborn
scikit-learn
psycopg2
sqlalchemy
python-dotenv
skyfield
tqdm
requests
